refactor: converted remove islands to ts
