--- src/challenges/remove-islands/index.test.js
@@ -1,27 +1,0 @@
-import remove_islands from './index';
-
-describe('YouTube: Remove Islands Challenge', () => {
-
-	test('should remove all the islands on the inside', () => {
-		let matrix = [
-			[1, 0, 0, 0, 0, 0],
-			[0, 1, 0, 1, 1, 1],
-			[0, 0, 1, 0, 1, 0],
-			[1, 1, 0, 0, 1, 0],
-			[1, 0, 1, 1, 0, 0],
-			[1, 0, 0, 0, 0, 1],
-		];
-
-		let expected = [
-			[1, 0, 0, 0, 0, 0],
-			[0, 0, 0, 1, 1, 1],
-			[0, 0, 0, 0, 1, 0],
-			[1, 1, 0, 0, 1, 0],
-			[1, 0, 0, 0, 0, 0],
-			[1, 0, 0, 0, 0, 1]
-		];
-
-		expect(remove_islands(matrix)).toEqual(expected);
-	});
-
-});
