+++ src/challenges/remove-islands/index.test.ts
@@ -1,0 +1,29 @@
+import { describe, expect, test } from 'bun:test';
+
+import remove_islands from './index';
+
+describe('YouTube: Remove Islands Challenge', () => {
+
+  test('should remove all the islands on the inside', () => {
+    let matrix = [
+      [1, 0, 0, 0, 0, 0],
+      [0, 1, 0, 1, 1, 1],
+      [0, 0, 1, 0, 1, 0],
+      [1, 1, 0, 0, 1, 0],
+      [1, 0, 1, 1, 0, 0],
+      [1, 0, 0, 0, 0, 1],
+    ];
+
+    let expected = [
+      [1, 0, 0, 0, 0, 0],
+      [0, 0, 0, 1, 1, 1],
+      [0, 0, 0, 0, 1, 0],
+      [1, 1, 0, 0, 1, 0],
+      [1, 0, 0, 0, 0, 0],
+      [1, 0, 0, 0, 0, 1]
+    ];
+
+    expect(remove_islands(matrix)).toEqual(expected);
+  });
+
+});
